Simpler React.createContext and React.useState objects.
Note: React.useContext dependency on the React.useEffect, React.useEffectLayout and React.useCallback
did not effect the behavior of these React Hooks. The React.useState does effect the behavior. This
is not entirely my expectation.

Not sure how propogation and state management are implemented with React v19 https://react.dev.